+++ src/v5/server.rs
@@ -11,17 +11,17 @@ use crate::service::{FactoryBuilder, FactoryBuilder2};
 use crate::{io::IoState, types::QoS};
 
 use super::codec as mqtt;
-use super::connect::{Connect, ConnectAck};
 use super::control::{ControlMessage, ControlResult};
 use super::default::{DefaultControlService, DefaultPublishService};
 use super::dispatcher::factory;
+use super::handshake::{Handshake, HandshakeAck};
 use super::publish::{Publish, PublishAck};
 use super::sink::{MqttSink, MqttSinkPool};
 use super::Session;
 
 /// Mqtt Server
 pub struct MqttServer<Io, St, C: ServiceFactory, Cn: ServiceFactory, P: ServiceFactory> {
-    connect: C,
+    handshake: C,
     srv_control: Cn,
     srv_publish: P,
     max_size: u32,
@@ -44,17 +44,17 @@ impl<Io, St, C>
     >
 where
     St: 'static,
-    C: ServiceFactory<Config = (), Request = Connect<Io>, Response = ConnectAck<Io, St>>
+    C: ServiceFactory<Config = (), Request = Handshake<Io>, Response = HandshakeAck<Io, St>>
         + 'static,
     C::Error: fmt::Debug,
 {
-    /// Create server factory and provide connect service
-    pub fn new<F>(connect: F) -> Self
+    /// Create server factory and provide handshake service
+    pub fn new<F>(handshake: F) -> Self
     where
         F: IntoServiceFactory<C>,
     {
         MqttServer {
-            connect: connect.into_factory(),
+            handshake: handshake.into_factory(),
             srv_control: DefaultControlService::default(),
             srv_publish: DefaultPublishService::default(),
             max_size: 0,
@@ -73,7 +73,7 @@ impl<Io, St, C, Cn, P> MqttServer<Io, St, C, Cn, P>
 where
     Io: AsyncRead + AsyncWrite + Unpin + 'static,
     St: 'static,
-    C: ServiceFactory<Config = (), Request = Connect<Io>, Response = ConnectAck<Io, St>>
+    C: ServiceFactory<Config = (), Request = Handshake<Io>, Response = HandshakeAck<Io, St>>
         + 'static,
     C::Error: fmt::Debug,
     Cn: ServiceFactory<
@@ -151,7 +151,7 @@ where
         C::Error: From<Srv::Error> + From<Srv::InitError>,
     {
         MqttServer {
-            connect: self.connect,
+            handshake: self.handshake,
             srv_publish: self.srv_publish,
             srv_control: service.into_factory(),
             max_size: self.max_size,
@@ -176,7 +176,7 @@ where
         PublishAck: TryFrom<Srv::Error, Error = C::Error>,
     {
         MqttServer {
-            connect: self.connect,
+            handshake: self.handshake,
             srv_publish: publish.into_factory(),
             srv_control: self.srv_control,
             max_size: self.max_size,
@@ -195,7 +195,7 @@ impl<Io, St, C, Cn, P> MqttServer<Io, St, C, Cn, P>
 where
     Io: AsyncRead + AsyncWrite + Unpin + 'static,
     St: 'static,
-    C: ServiceFactory<Config = (), Request = Connect<Io>, Response = ConnectAck<Io, St>>
+    C: ServiceFactory<Config = (), Request = Handshake<Io>, Response = HandshakeAck<Io, St>>
         + 'static,
     C::Error: From<Cn::Error>
         + From<Cn::InitError>
@@ -216,7 +216,7 @@ where
         self,
     ) -> impl ServiceFactory<Config = (), Request = Io, Response = (), Error = MqttError<C::Error>>
     {
-        let connect = self.connect;
+        let handshake = self.handshake;
         let publish = self.srv_publish.map_init_err(|e| MqttError::Service(e.into()));
         let control = self
             .srv_control
@@ -225,7 +225,7 @@ where
 
         ntex::unit_config(
             FactoryBuilder::new(handshake_service_factory(
-                connect,
+                handshake,
                 self.max_size,
                 self.max_receive,
                 self.max_topic_alias,
@@ -248,7 +248,7 @@ where
         Error = MqttError<C::Error>,
         InitError = C::InitError,
     > {
-        let connect = self.connect;
+        let handshake = self.handshake;
         let publish = self.srv_publish.map_init_err(|e| MqttError::Service(e.into()));
         let control = self
             .srv_control
@@ -257,7 +257,7 @@ where
 
         ntex::unit_config(
             FactoryBuilder2::new(handshake_service_factory2(
-                connect,
+                handshake,
                 self.max_size,
                 self.max_receive,
                 self.max_topic_alias,
@@ -287,7 +287,7 @@ fn handshake_service_factory<Io, St, C>(
 >
 where
     Io: AsyncRead + AsyncWrite + Unpin,
-    C: ServiceFactory<Config = (), Request = Connect<Io>, Response = ConnectAck<Io, St>>,
+    C: ServiceFactory<Config = (), Request = Handshake<Io>, Response = HandshakeAck<Io, St>>,
     C::Error: fmt::Debug,
 {
     ntex::apply(
@@ -336,7 +336,7 @@ fn handshake_service_factory2<Io, St, C>(
 >
 where
     Io: AsyncRead + AsyncWrite + Unpin,
-    C: ServiceFactory<Config = (), Request = Connect<Io>, Response = ConnectAck<Io, St>>,
+    C: ServiceFactory<Config = (), Request = Handshake<Io>, Response = HandshakeAck<Io, St>>,
     C::Error: fmt::Debug,
 {
     ntex::apply(
@@ -380,7 +380,7 @@ async fn handshake<Io, S, St, E>(
 ) -> Result<(Io, IoState<mqtt::Codec>, Session<St>, u16), S::Error>
 where
     Io: AsyncRead + AsyncWrite + Unpin,
-    S: Service<Request = Connect<Io>, Response = ConnectAck<Io, St>, Error = MqttError<E>>,
+    S: Service<Request = Handshake<Io>, Response = HandshakeAck<Io, St>, Error = MqttError<E>>,
 {
     log::trace!("Starting mqtt v5 handshake");
 
@@ -420,7 +420,7 @@ where
 
             // authenticate mqtt connection
             let mut ack = service
-                .call(Connect::new(
+                .call(Handshake::new(
                     connect,
                     io,
                     sink,